requirements.txt
altair
pandas
streamlit
snowflake-connector-python==2.7.6
matplotlib